Fboemer/clang tidy (#289)
Fixed most warnings in
```bash
python /usr/bin/run-clang-tidy-6.0.py \
-checks="-*,\
hicpp-*,\
-hicpp-no-array-decay,\
-hicpp-special-member-functions,\
-hicpp-vararg*,\
llvm-*,\
-llvm-include-order,\
misc-*,\
modernize-*,\
performance-*,\
readability-*,\
"
```

+++ test/test_server_client.in.cpp
@@ -482,7 +482,8 @@ auto server_client_relu_packed_test = [](size_t element_count,
   vector<float> inputs(shape_size(shape));
   vector<float> exp_results(shape_size(shape));
   for (size_t i = 0; i < shape_size(shape); ++i) {
-    inputs[i] = static_cast<int>(i) - static_cast<int>(shape_size(shape)) / 2;
+    inputs[i] =
+        static_cast<float>(i) - static_cast<float>(shape_size(shape)) / 2.0;
 
     if (bounded) {
       exp_results[i] = bounded_relu(inputs[i]);
